+++ apps/src/componentLibrary/button/Button.story.tsx
@@ -112,6 +112,13 @@ GroupOfColorsOfButtons.args = {
       onClick: () => null,
     },
     {
+      text: 'Button Secondary Grey',
+      color: buttonColors.grey,
+      type: 'secondary',
+      size: 'm',
+      onClick: () => null,
+    },
+    {
       text: 'Button Secondary White',
       color: buttonColors.white,
       type: 'secondary',
@@ -174,6 +181,14 @@ GroupOfColorsOfButtons.args = {
     {
       icon: {iconName: 'smile', iconStyle: 'solid'},
       color: buttonColors.black,
+      type: 'secondary',
+      isIconOnly: true,
+      size: 'm',
+      onClick: () => null,
+    },
+    {
+      icon: {iconName: 'smile', iconStyle: 'solid'},
+      color: buttonColors.grey,
       type: 'secondary',
       isIconOnly: true,
       size: 'm',